Refatoração final
