+++ semana-14/projeto/src/utils/index.ts
@@ -1,0 +1,4 @@
+export { default as getClients } from './getClients';
+export { default as validateFullName } from './validateFullName';
+export { default as validateCPF } from './validateCPF';
+export { default as validateBirthDate } from './validadeBirthDate';
